+++ Tests/Desktop/Bugs/GH227.cs
@@ -46,6 +46,8 @@ namespace Dicom.Bugs
                 stream.Seek(0, SeekOrigin.Begin);
 
                 var file = DicomFile.Open(stream);
+                Assert.Equal(DicomTransferSyntax.DeflatedExplicitVRLittleEndian, file.Dataset.InternalTransferSyntax);
+
                 const int expected = 16;
                 var actual = file.Dataset.Get<int>(DicomTag.BitsAllocated);
                 Assert.Equal(expected, actual);
@@ -65,8 +67,28 @@ namespace Dicom.Bugs
                 stream.Seek(0, SeekOrigin.Begin);
 
                 var file = DicomFile.Open(stream);
+                Assert.Equal(DicomTransferSyntax.DeflatedExplicitVRLittleEndian, file.Dataset.InternalTransferSyntax);
+
                 const int expected = 16;
                 var actual = file.Dataset.Get<int>(DicomTag.BitsAllocated);
+                Assert.Equal(expected, actual);
+            }
+        }
+
+        [Fact]
+        public void Save_DeflatedFile_CorrectlyCompressed()
+        {
+            var deflated = DicomFile.Open(@"Test Data\GH227.dcm");
+
+            using (var stream = new MemoryStream())
+            {
+                deflated.Save(stream);
+                File.WriteAllBytes("GH227_dfldump.dcm", stream.ToArray());
+                stream.Seek(0, SeekOrigin.Begin);
+
+                var file = DicomFile.Open(stream);
+                const int expected = 512;
+                var actual = file.Dataset.Get<int>(DicomTag.Columns);
                 Assert.Equal(expected, actual);
             }
         }